requirements.txt
# Django et extensions principales
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.4
django-crispy-forms==2.1
crispy-bootstrap5==0.7

# Base de données
psycopg2-binary==2.9.9
django-environ==0.11.2

# Authentification et sécurité
django-allauth==0.57.0
django-guardian==2.4.0
django-ratelimit==4.1.0

# Cache et performance
redis==5.0.1
django-redis==5.4.0
django-cachalot==2.6.1

# Tâches asynchrones
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Images et fichiers
Pillow==10.1.0
django-imagekit==5.0.0
python-magic==0.4.27

# API et sérialisation
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
django-rest-auth==0.9.5

# Utilitaires
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0

# Pagination et recherche
django-haystack==3.2.1
elasticsearch==8.11.0
django-elasticsearch-dsl==7.3

# Internationalisation
django-modeltranslation==0.18.11
babel==2.13.1

# Monitoring et logs
django-extensions==3.2.3
django-debug-toolbar==4.2.0
sentry-sdk==1.38.0

# Tests
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0

# Déploiement
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.34.0

# Développement
black==23.11.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.6.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Utilitaires système
psutil==5.9.6
python-dateutil==2.8.2
pytz==2023.3

# Email
django-anymail==10.2

# Export et rapports
reportlab==4.0.7
openpyxl==3.1.2
xlsxwriter==3.1.9

# Validation et nettoyage
bleach==6.1.0
html5lib==1.1

# Géolocalisation (pour les livraisons)
geopy==2.4.1
django-location-field==2.7.0

# QR Codes (pour les livres)
qrcode==7.4.2
python-barcode==0.15.1

# Notifications
django-notifications-hq==1.8.3
channels==4.0.0
channels-redis==4.1.0

# Backup et maintenance
django-dbbackup==4.0.2
django-cleanup==8.0.0

# Sécurité avancée
django-csp==3.7
django-permissions-policy==4.17.0
